Private types and functions

Documentation for FourierFlows.jl's internal interface.

FourierFlows

FourierFlows.AbstractDiagnostic — Type

Abstract supertype for diagnostics.

Problem

FourierFlows.Clock — Type

Clock{T<:AbstractFloat}

A struct containing the time-step dt, the time t and the step-number of the simulation.

FourierFlows.EmptyParams — Type

EmptyParams <: AbstractParams

A placeholder struct for parameters.

FourierFlows.EmptyVars — Type

EmptyVars <: AbstractVars

A placeholder struct for variables.

FourierFlows.Equation — Type

Equation{T, TL, G<:AbstractFloat}

A struct that includes the equation to be solved ∂u/∂t = L*u + N(u). Array L includes the coefficients of the linear term L*u and calcN! is a function which computes the nonlinear term N(u). The struct also includes the problem's grid and the float type of the state vector (and consequently of N(u)).

FourierFlows.Equation — Method

Equation(L, calcN!, grid; dims=supersize(L), T=nothing)

An equation constructor given the array L of the coefficients of the linear term, the function calcN! that computes the nonlinear term and the problem's grid.

FourierFlows.Problem — Type

Problem{T, A<:AbstractArray, Tg<:AbstractFloat, TL}

A struct including everything a FourierFlows problem requires: the state vector sol, the clock, the equation eqn, the grid, all problem variables in vars, problem parameters in params, and the timestepper.

FourierFlows.Problem — Method

Problem(eqn::Equation, stepper, dt, grid::AbstractGrid,
        vars=EmptyVars, params=EmptyParams, dev::Device=CPU(); stepperkwargs...)

Construct a Problem for equation eqn using the timestepper with timestep dt, on grid and device. Optionally provide variables in vars and parameters with params. The stepperkwargs are passed to the time-stepper constructor.

Domain

FourierFlows.getaliasedwavenumbers — Method

getaliasedwavenumbers(nk, nkr, aliasfraction)

Returns the top aliasfraction highest wavenumbers.

FourierFlows.griddevice — Method

griddevice(grid)

Returns the device on which the grid lives on.

FourierFlows.makefilter — Method

makefilter(K; order=4, innerK=0.65, outerK=1)

Returns a filter acting on the non-dimensional wavenumber K that decays exponentially for K>innerK, thus removing high-wavenumber content from a spectrum it is multiplied with. The decay rate is determined by order and outerK determines the outer wavenumber at which the filter is smaller than Float64 machine precision.

Utilities

FourierFlows.getfieldspecs — Method

getfieldspecs(fieldnames, fieldtype)

Returns an array of (fieldname[i], fieldtype) tuples.

FourierFlows.jacobian — Method

jacobian(a, b, g)

Returns the Jacobian of a and b.

FourierFlows.jacobianh — Method

jacobianh(a, b, g)

Returns the transform of the Jacobian of two fields a, b on the grid g.

FourierFlows.parsevalsum — Method

parsevalsum(uh, g)

Returns real(Σ uh) on the grid g.

FourierFlows.parsevalsum2 — Method

parsevalsum2(uh, g)

Returns ∫|u|² = Σ|uh|² on the grid g, where uh is the Fourier transform of u.

FourierFlows.radialspectrum — Method

radialspectrum(ah, g; n=nothing, m=nothing, refinement=2)

Returns aρ = ∫ ah(ρ,θ) ρ dρ dθ, the radial spectrum of ah known on the Cartesian wavenumber grid (k,l).

aρ is found by intepolating ah onto a polar wavenumber grid (ρ,θ), and then integrating over θ to find aρ. The default resolution (n,m) for the polar wave number grid is n=refinement*maximum(nk, nl), m=refinement*maximum(nk, nl), where refinement=2 by default. If ah is in conjugate symmetric form only the upper half plane in θ is represented on the polar grid.

FourierFlows.varsexpression — Method

varsexpression(name, fieldspecs; parent=:AbstractVars, typeparams=nothing)

varsexpression(name, fieldspecs...; parent=:AbstractVars, typeparams=nothing)

varsexpression(name, physicalfields, fourierfields; physicaltype=:Tp, fouriertype=:Tf,
               parent=:AbstractVars, typeparams=[physicaltype, fouriertype])

Returns an expression that defines an AbstractVars type.

Diagnostics

Base.getindex — Method

e.g. plot(energydiag[:t], energydiag[:data]).

FourierFlows.extend! — Method

extend!(diag::AbstractDiagnostic, n)

Extend the data, time, and steps vectors of diag by n.

Output

FourierFlows.savefield — Method

savefield(file, location, data)

Saves a particular field.

FourierFlows.savefields — Method

savefields(file, field)

Saves some parameters of prob.field.

FourierFlows.uniquepath — Method

uniquepath(path)

Returns path with a number appended if isfile(path), incremented until path does not exist.

Timesteppers

FourierFlows.getetdcoeffs — Method

getetdcoeffs(dt, L; ncirc=32, rcirc=1)

Calculate ETDRK4 coefficients associated with the (diagonal) linear coefficient L by integrating over a small circle in complex space.

Diffusion Testbed Module

FourierFlows.Diffusion.Vars — Method

Vars(dev, grid)

Returns the variables vars for a constant diffusivity problem on grid.

FourierFlows.Diffusion.Equation — Method

Equation(dev, κ, grid)

Returns the equation for a constant diffusivity problem on grid with diffusivity κ.

FourierFlows.Diffusion.calcN! — Method

calcN!(N, sol, t, clock, vars, params, grid)

Calculate the nonlinear term for the 1D heat equation.