Private types and functions

Documentation for FourierFlows.jl's internal interface.

FourierFlows

FourierFlows.AbstractDiagnostic — Type

Abstract supertype for diagnostics.

Problem

FourierFlows.Clock — Type

Clock{T<:AbstractFloat}

A struct containing the time-step dt, the time t and the step-number of the simulation.

FourierFlows.EmptyParams — Type

EmptyParams <: AbstractParams

A placeholder struct for parameters.

FourierFlows.EmptyVars — Type

EmptyVars <: AbstractVars

A placeholder struct for variables.

FourierFlows.Equation — Type

struct Equation{T, TL, G<:AbstractFloat}

A struct that includes the equation to be solved ∂u/∂t = L*u + N(u). Array L includes the coefficients of the linear term L*u and calcN! is a function which computes the nonlinear term N(u). The struct also includes the problem's grid and the float type of the state vector (and consequently of N(u)).

L::Any

array with coefficient for the linear part of the equation

calcN!::Function

function that computes the nonlinear part of the equation

grid::AbstractGrid{G,A,Alias} where Alias where A where G<:AbstractFloat

the grid

dims::Tuple

the dimensions of L

T::Any

the float type for the state vector

FourierFlows.Equation — Method

Equation(L, calcN!, grid; dims=supersize(L), T=nothing)

The equation constructor from the array L of the coefficients of the linear term, the function calcN! that computes the nonlinear term and the grid for the problem.

FourierFlows.Problem — Type

struct Problem{T, A<:AbstractArray, Tg<:AbstractFloat, TL}

A struct including everything a FourierFlows problem requires: the state vector sol, the clock, the equation eqn, the grid, all problem variables in vars, problem parameters in params, and the timestepper.

FourierFlows.Problem — Method

Problem(eqn::Equation, stepper, dt, grid::AbstractGrid{T}, 
        vars=EmptyVars, params=EmptyParams, dev::Device=CPU(); stepperkwargs...) where T

Construct a Problem for equation eqn using the timestepper with timestep dt, on grid and device. Optionally provide variables in vars and parameters with params. The stepperkwargs are passed to the time-stepper constructor.

Domain

FourierFlows.getaliasedwavenumbers — Method

getaliasedwavenumbers(nk, nkr, aliased_fraction)

Returns the top aliased_fraction highest wavenumbers, both for and real FFTs, kalias and kralias respectively. For example, aliased_fraction=1/3 should return the indices of the top-most 1/6-th (in absolute value) for both positive and negative wavenumbers (i.e., 1/3 total) that should be set to zero after performing an FFT.

FourierFlows.griddevice — Method

griddevice(grid)

Returns the device on which the grid lives on.

FourierFlows.makefilter — Method

makefilter(K; order=4, innerK=0.65, outerK=1)

Returns a filter acting on the non-dimensional wavenumber K that decays exponentially for K>innerK, thus removing high-wavenumber content from a spectrum it is multiplied with. The decay rate is determined by order and outerK determines the outer wavenumber at which the filter is smaller than Float64 machine precision.

Utilities

FourierFlows.jacobian — Method

jacobian(a, b, grid)

Returns the Jacobian of a and b on grid.

FourierFlows.jacobianh — Method

jacobianh(a, b, grid)

Returns the Fourier transform of the Jacobian of a and b on grid.

FourierFlows.on_grid — Method

on_grid(func, grid)

Returns an array, of the ArrayType of the device grid lives on, that contains the values of function func evaluated on the grid.

FourierFlows.parsevalsum — Method

parsevalsum(uh, grid)

Returns real(Σ uh) on the grid, i.e.

\[ℜ [ \sum_{𝐤} û_{𝐤} L_x L_y ] \,,\]

where $û_{𝐤} =$ uh $/($ grid.nx $e^{- i 𝐤 ⋅ 𝐱₀})$, with $𝐱₀$ the vector with components the left-most position in each direction.

FourierFlows.parsevalsum2 — Method

parsevalsum2(uh, grid)

Returns Σ |uh|² on the grid, which is equal to the domain integral of u. More specifically, it returns

\[\sum_{𝐤} |û_{𝐤}|² L_x L_y = \int u(𝐱)² \, 𝖽x 𝖽y \,,\]

where $û_{𝐤} =$ uh $/($ grid.nx $e^{- i 𝐤 ⋅ 𝐱₀})$, with $𝐱₀$ the vector with components the left-most position in each direction.

FourierFlows.radialspectrum — Method

radialspectrum(fh, grid; n=nothing, m=nothing, refinement=2)

Returns the radial spectrum of fh. fh lives on Cartesian wavenumber grid $(k, l)$. To compute the radial spectrum, we first interpolate $f̂(k, l)$ onto a radial wavenumber grid $(ρ, θ)$, where $ρ² = k²+l²$ and $θ = \arctan(l/k)$. Note here that $f̂ =$ fh $/($ grid.nx $e^{- i 𝐤 ⋅ 𝐱₀})$, with $𝐱₀$ the vector with components the left-most position in each direction. After interpolation, we integrate $f̂$over angles $θ$ to get fρ,

\[ f̂_ρ = \int f̂(ρ, θ) ρ 𝖽ρ 𝖽θ \, .\]

The resolution (n, m) for the polar wavenumber grid is n = refinement * maximum(nk, nl), m = refinement * maximum(nk, nl), where refinement = 2 by default. If fh is in conjugate symmetric form then only the upper-half plane in $θ$ is represented on the polar grid.

Diagnostics

Base.getindex — Method

e.g. plot(energydiag[:t], energydiag[:data]).

FourierFlows.extend! — Method

extend!(diag::AbstractDiagnostic, n)

Extend the data, time, and steps vectors of diag by n.

Output

FourierFlows.savefield — Method

savefield(file, location, data)

Saves a particular field.

FourierFlows.savefields — Method

savefields(file, field)

Saves some parameters of prob.field.

FourierFlows.uniquepath — Method

uniquepath(path)

Returns path with a number appended if isfile(path), incremented until path does not exist.

Timesteppers

FourierFlows.getetdcoeffs — Method

getetdcoeffs(dt, L; ncirc=32, rcirc=1)

Calculate ETDRK4 coefficients associated with the (diagonal) linear coefficient L by integrating over a small circle in complex space.

Diffusion Testbed Module

FourierFlows.Diffusion.Params — Type

struct Params{T} <: AbstractParams

The parameters for diffusion problem:

κ::Any

diffusivity coefficient

FourierFlows.Diffusion.Vars — Type

struct Vars{Aphys, Atrans} <: AbstractVars

The variables for diffusion problem:

c

tracer concentration

cx

tracer concentration derivative ∂ₓc

ch

Fourier transform of tracer concentration

cxh

Fourier transform of tracer concentration derivative ∂ₓc

FourierFlows.Diffusion.Vars — Method

Vars(dev, grid)

Returns the variables vars for a constant diffusivity problem on grid.

FourierFlows.Diffusion.Equation — Method

Equation(dev, κ, grid)

Returns the equation for a constant diffusivity problem on grid with diffusivity κ.

FourierFlows.Diffusion.calcN! — Method

calcN!(N, sol, t, clock, vars, params, grid)

Calculate the nonlinear term for the 1D diffusion equation.