Public Documentation

Documentation for FourierFlows.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

FourierFlows

FourierFlows.FourierFlows — Module

Main module for FourierFlows.jl – an ecosystem for solving partial differential equations on periodic domains using Fourier-based pseudospectral methods."

Exports

AB3TimeStepper

AS

AbstractGrid

AbstractParams

AbstractTimeStepper

AbstractVars

AnyCuArray

AnyCuMatrix

AnyCuVecOrMat

AnyCuVector

ArrayType

CPU

CUA_NULL

CUBLAS

CUDA

CUDNN

CUFFT

CUPTI

CURAND

CUSOLVER

CUSPARSE

CUTENSOR

CU_NULL

CuArray

CuArrayPtr

CuContext

CuCurrentContext

CuCurrentDevice

CuDefaultStream

CuDevice

CuDeviceArray

CuDeviceMatrix

CuDeviceStream

CuDeviceTexture

CuDeviceVector

CuDim

CuDim3

CuError

CuEvent

CuFunction

CuGlobal

CuIterator

CuLink

CuLinkImage

CuMatrix

CuModule

CuModuleFile

CuPrimaryContext

CuPtr

CuRef

CuStream

CuStreamLegacy

CuStreamPerThread

CuTexture

CuTextureArray

CuVecOrMat

CuVector

DenseCuArray

DenseCuMatrix

DenseCuVecOrMat

DenseCuVector

Device

Diagnostic

ETDRK4TimeStepper

FULL_MASK

FilteredAB3TimeStepper

FilteredETDRK4TimeStepper

FilteredForwardEulerTimeStepper

FilteredRK4TimeStepper

ForwardEulerTimeStepper

GPU

Mem

NVML

NVTX

OneDGrid

OutOfGPUMemoryError

Output

PtrOrCuPtr

RK4TimeStepper

RefOrCuRef

StridedCuArray

StridedCuMatrix

StridedCuVecOrMat

StridedCuVector

ThreeDGrid

TimeStepper

TwoDGrid

WMMA

ZeroDGrid

activate

active_blocks

add_data!

add_file!

@atomic

attribute

attribute!

attributes

blockDim

blockIdx

cache_config

cache_config!

capability

clock

complete

context

context!

@createarrays

cu

@cuDynamicSharedMem

@cuStaticSharedMem

@cuassert

@cuda

cudacall

cudaconvert

cufunction

@cuprint

@cuprintf

@cuprintln

@cushow

cxtype

dealias!

device

device!

@device_code

@device_code_llvm

@device_code_lowered

@device_code_ptx

@device_code_sass

@device_code_typed

@device_code_warntype

device_reset!

deviceid

devices

@devzeros

devzeros

dynamic_cufunction

elapsed

fft

flags

fltype

get

gridDim

gridpoints

groupsize

has_cuda

has_cuda_gpu

has_cudnn

has_cupti

has_cutensor

has_nvml

has_nvtx

ifft

increment!

innereltype

irfft

is_managed

isactive

launch_configuration

ldg

limit

limit!

malloc

memory_type

name

nanosleep

ndevices

nextwarp

occupancy

prevwarp

priority

priority_range

rand_logn!

rand_poisson!

record

resize!

rfft

savediagnostic

saveoutput

saveproblem

set

setflags!

shfl_down_sync

shfl_sync

shfl_up_sync

shfl_xor_sync

shmem_config

shmem_config!

step_until!

stepforward!

supersize

@superzeros

superzeros

sync_grid

sync_threads

sync_threads_and

sync_threads_count

sync_threads_or

sync_warp

synchronize

this_grid

threadIdx

threadfence

threadfence_block

threadfence_system

totalmem

unsafe_cached_load

unsafe_reset!

update!

uuid

vote_all

vote_any

vote_ballot

warpsize

@zeros

FourierFlows.AbstractGrid — Type

Abstract supertype for grids.

FourierFlows.AbstractParams — Type

Abstract supertype for parameters.

FourierFlows.AbstractTimeStepper — Type

Abstract supertype for timesteppers.

FourierFlows.AbstractVars — Type

Abstract supertype for variables.

FourierFlows.CPU — Type

CPU device.

FourierFlows.Device — Type

Abstract supertype for device.

FourierFlows.GPU — Type

GPU device.

Problem

Domain

FourierFlows.OneDGrid — Type

struct OneDGrid{T<:AbstractFloat, Tk, Tx, Tfft, Trfft, Talias} <: AbstractGrid{T, Tk, Talias}

A one-dimensional grid.

nx::Int64

number of points in x

nk::Int64

number of wavenumbers in x

nkr::Int64

number of positive wavenumbers in x (real Fourier transforms)

dx::AbstractFloat

grid spacing in x

Lx::AbstractFloat

domain extent in x

x::Any

range with x-grid-points

k::Any

array with x-wavenumbers

kr::Any

array with positive x-wavenumbers (real Fourier transforms)

invksq::Any

array with inverse squared k-wavenumbers, 1/k²

invkrsq::Any

array with inverse squared kr-wavenumbers, 1/kr²

fftplan::Any

the FFT plan for complex-valued fields

rfftplan::Any

the FFT plan for real-valued fields

aliased_fraction::AbstractFloat

the fraction of wavenumbers that are aliased (e.g., 1/3 for quadradic nonlinearities)

kalias::Any

range of the indices of aliased x-wavenumbers

kralias::Any

range of the indices of aliased positive x-wavenumbers (real Fourier transforms)

FourierFlows.OneDGrid — Method

OneDGrid(nx, Lx; x0=-Lx/2, nthreads=Sys.CPU_THREADS, effort=FFTW.MEASURE, 
                  T=Float64, aliased_fraction=1/3, ArrayType=Array)

Constructs a OneDGrid object with size Lx, resolution nx, and leftmost position x0. FFT plans are generated for nthreads CPUs using FFTW flag effort. The float type is T and the array types is ArrayType. The aliased_fraction keyword determines the highest wavenubers that are being zero-ed out by dealias! function; 1/3 is the nominal value for quadratic nonlinearities.

FourierFlows.ThreeDGrid — Type

struct ThreeDGrid{T<:AbstractFloat, Tk, Tx, Tfft, Trfft, Talias} <: AbstractGrid{T, Tk, Talias}

A three-dimensional grid.

nx::Int64

number of points in x

ny::Int64

number of points in y

nz::Int64

number of points in z

nk::Int64

number of wavenumbers in x

nl::Int64

number of wavenumbers in y

nm::Int64

number of wavenumbers in z

nkr::Int64

number of positive wavenumers in x (real Fourier transforms)

dx::AbstractFloat

grid spacing in x

dy::AbstractFloat

grid spacing in y

dz::AbstractFloat

grid spacing in z

Lx::AbstractFloat

domain extent in x

Ly::AbstractFloat

domain extent in y

Lz::AbstractFloat

domain extent in z

x::Any

range with x-grid-points

y::Any

range with y-grid-points

z::Any

range with z-grid-points

k::Any

array with x-wavenumbers

l::Any

array with y-wavenumbers

m::Any

array with z-wavenumbers

kr::Any

array with positive x-wavenumbers (real Fourier transforms)

Ksq::Any

array with squared total wavenumbers, k²+l²+m²

invKsq::Any

array with inverse squared total wavenumbers, 1/(k²+l²+m²)

Krsq::Any

array with squared total wavenumbers for real Fourier transforms, kr²+l²+m²

invKrsq::Any

array with inverse squared total wavenumbers for real Fourier transforms, 1/(kr²+l²+m²)

fftplan::Any

the FFT plan for complex-valued fields

rfftplan::Any

the FFT plan for real-valued fields

aliased_fraction::AbstractFloat

the fraction of wavenumbers that are aliased (e.g., 1/3 for quadradic nonlinearities)

kalias::Any

range of the indices of aliased x-wavenumbers

kralias::Any

range of the indices of aliased positive x-wavenumbers (real Fourier transforms)

lalias::Any

range of the indices of aliased y-wavenumbers

malias::Any

range of the indices of aliased y-wavenumbers

FourierFlows.ThreeDGrid — Type

ThreeDGrid(nx, Lx, ny=nx, Ly=Lx, nz=nx, Lz=Lx; x0=-Lx/2, y0=-Ly/2, z0=-Lz/2, 
nthreads=Sys.CPU_THREADS, effort=FFTW.MEASURE, T=Float64, aliased_fraction=1/3, ArrayType=Array)

Constructs a TwoDGrid object with size Lx, Ly, Lz, resolution nx, ny, nz and leftmost positions x0, y0, z0. FFT plans are generated for nthreads CPUs using FFTW flag effort. The float type is T and the array types is ArrayType. The aliased_fraction keyword determines the highest wavenubers that are being zero-ed out by dealias! function; 1/3 is the nominal value for quadratic nonlinearities.

FourierFlows.TwoDGrid — Type

struct TwoDGrid{T<:AbstractFloat, Tk, Tx, Tfft, Trfft, Talias} <: AbstractGrid{T, Tk, Talias}

A two-dimensional grid.

nx::Int64

number of points in x

ny::Int64

number of points in y

nk::Int64

number of wavenumbers in x

nl::Int64

number of wavenumbers in y

nkr::Int64

number of positive wavenumers in x (real Fourier transforms)

dx::AbstractFloat

grid spacing in x

dy::AbstractFloat

grid spacing in y

Lx::AbstractFloat

domain extent in x

Ly::AbstractFloat

domain extent in y

x::Any

range with x-grid-points

y::Any

range with y-grid-points

k::Any

array with x-wavenumbers

l::Any

array with y-wavenumbers

kr::Any

array with positive x-wavenumbers (real Fourier transforms)

Ksq::Any

array with squared total wavenumbers, k²+l²

invKsq::Any

array with inverse squared total wavenumbers, 1/(k²+l²)

Krsq::Any

array with squared total wavenumbers for real Fourier transforms, kr²+l²

invKrsq::Any

array with inverse squared total wavenumbers for real Fourier transforms, 1/(kr²+l²)

fftplan::Any

the FFT plan for complex-valued fields

rfftplan::Any

the FFT plan for real-valued fields

aliased_fraction::AbstractFloat

the fraction of wavenumbers that are aliased (e.g., 1/3 for quadradic nonlinearities)

kalias::Any

range of the indices of aliased x-wavenumbers

kralias::Any

range of the indices of aliased positive x-wavenumbers (real Fourier transforms)

lalias::Any

range of the indices of aliased y-wavenumbers

FourierFlows.TwoDGrid — Type

TwoDGrid(nx, Lx, ny=nx, Ly=Lx; x0=-Lx/2, y0=-Ly/2, nthreads=Sys.CPU_THREADS, 
                  effort=FFTW.MEASURE, T=Float64, aliased_fraction=1/3, ArrayType=Array)

Constructs a TwoDGrid object with size Lx, Ly, resolution nx, ny, and leftmost positions x0, y0. FFT plans are generated for nthreads CPUs using FFTW flag effort. The float type is T and the array types is ArrayType. The aliased_fraction keyword determines the highest wavenubers that are being zero-ed out by dealias! function; 1/3 is the nominal value for quadratic nonlinearities.

FourierFlows.ZeroDGrid — Type

A placeholder grid object for 0D problems (in other words, systems of ODEs).

FourierFlows.dealias! — Method

dealias!(a, grid, kalias)

Dealias array a on the grid with aliased x-wavenumbers kalias.

FourierFlows.gridpoints — Method

gridpoints(grid::TwoDGrid)
gridpoints(grid::ThreeDGrid)

Returns the collocation points of the grid in 2D or 3D arrays X, Y (and Z).

Utilities

FourierFlows.ArrayType — Method

ArrayType(::Device)
ArrayType(::Device, T, dim)

Returns the proper array type according to the Device chosen, i.e., Array for CPU and CuArray for GPU.

FourierFlows.cxtype — Method

cxtype(T)

Returns T when T is Complex, or Complex{T} when T is Real.

FourierFlows.devzeros — Method

devzeros(dev, T, dims)

Returns an array like A of type T, but full of zeros.

FourierFlows.fltype — Method

fltype(T)

Returns T when T<:AbstractFloat or Tf when T<:Complex{Tf}.

FourierFlows.innereltype — Method

innereltype(x)

Recursively determine the 'innermost' type in by the collection x (which may be, for example, a collection of a collection).

FourierFlows.superzeros — Method

superzeros(T, A)

Returns an array like A, but full of zeros. If innereltype(A) can be promoted to T, then the innermost elements of the array will have type T.

FourierFlows.@devzeros — Macro

@devzeros dev T dims a b c...

Create arrays of all zeros with element type T, size dims, and global names a, b, c (for example) on device dev.

FourierFlows.@superzeros — Macro

@superzeros T a b c d...
@superzeros T dims b c d...

Generate arrays b, c, d... with the super-dimensions of a and innereltype T.

FourierFlows.@zeros — Macro

@zeros T dims a b c...

Create arrays of all zeros with element type T, size dims, and global names a, b, c (for example). An arbitrary number of arrays may be created.

Diagnostics

FourierFlows.Diagnostic — Type

Diagnostic(calc, prob; freq=1, nsteps=100, N=floor(Int, (nsteps+1)/freq))

Construct a diagnostic which stores the result of calc(prob) with frequency freq

FourierFlows.increment! — Method

increment!(diag)

Increment the Diagnostic diag, or an array of Diagnostics diags.

FourierFlows.update! — Method

update!(diag)

Update diag.

Output

FourierFlows.Output — Type

Output(prob, filename, fieldtuples...)

Define output for prob with fields and functions that calculate the output in the list of tuples fieldtuples = (fldname, func)....

FourierFlows.savediagnostic — Method

savediagnostic(diag, diagname)

Save diagnostics in diag to file, labeled by diagname.

FourierFlows.saveoutput — Method

saveoutput(out)

Save the fields in out.fields to out.path.

FourierFlows.saveproblem — Method

saveproblem(prob, filename)

Save certain aspects of a problem.

Timesteppers

FourierFlows.AB3TimeStepper — Type

AB3TimeStepper{T} <: AbstractTimeStepper{T}

Type for 3rd order Adams-Bashforth time-stepper.

FourierFlows.AB3TimeStepper — Type

FilteredAB3TimeStepper(equation, dev; filterkwargs...)

Construct a 3rd order Adams-Bashforth time stepper.

FourierFlows.ETDRK4TimeStepper — Type

ETDRK4TimeStepper{T,TL} <: AbstractTimeStepper{T}

Type for 4th-order exponential-time-differencing Runge-Kutta time stepper.

FourierFlows.ETDRK4TimeStepper — Type

ETDRK4TimeStepper(eq::Equation, dt, dev::Device=CPU())

Construct a 4th-order exponential-time-differencing Runge-Kutta time stepper.

FourierFlows.FilteredAB3TimeStepper — Type

FilteredAB3TimeStepper{T} <: AbstractTimeStepper{T}

Type for 3rd order Adams-Bashforth time-stepper with spectral filtering.

FourierFlows.FilteredAB3TimeStepper — Type

FilteredAB3TimeStepper(equation, dev; filterkwargs...)

Construct a 3rd order Adams-Bashforth time stepper with spectral filtering.

FourierFlows.FilteredETDRK4TimeStepper — Type

FilteredETDRK4TimeStepper{T,TL,Tf} <: AbstractTimeStepper{T}

Type for 4th-order exponential-time-differencing Runge-Kutta time stepper with spectral filtering.

FourierFlows.FilteredETDRK4TimeStepper — Type

FilteredETDRK4TimeStepper(equation, dt; filterkwargs...)

Construct a 4th-order exponential-time-differencing Runge-Kutta time stepper with spectral filtering.

FourierFlows.FilteredForwardEulerTimeStepper — Type

FilteredForwardEulerTimeStepper(eq, dev; filterkwargs...)

Type for Forward Euler timestepper with spectral filtering.

FourierFlows.FilteredForwardEulerTimeStepper — Type

FilteredForwardEulerTimeStepper(eq, dev; filterkwargs...)

Construct a forward Euler timestepper with spectral filtering.

FourierFlows.FilteredRK4TimeStepper — Type

FilteredRK4TimeStepper{T,Tf} <: AbstractTimeStepper{T}

Type for 4th-order Runge-Kutta time stepper with spectral filtering for the equation eq.

FourierFlows.FilteredRK4TimeStepper — Type

FilteredRK4TimeStepper(eq::Equation, dev::Device=CPU(); filterkwargs...)

Construct a 4th-order Runge-Kutta time stepper with spectral filtering for the equation eq.

FourierFlows.ForwardEulerTimeStepper — Type

ForwardEulerTimeStepper{T} <: AbstractTimeStepper{T}

Type for Forward Euler timestepper.

FourierFlows.ForwardEulerTimeStepper — Type

ForwardEulerTimeStepper(eq::Equation, dev::Device=CPU())

Construct a Forward Euler timestepper.

FourierFlows.RK4TimeStepper — Type

RK4TimeStepper{T} <: AbstractTimeStepper{T}

Type for 4th-order Runge-Kutta time stepper for the equation eq.

FourierFlows.RK4TimeStepper — Type

RK4TimeStepper(eq::Equation, dev::Device=CPU())

Construct a 4th-order Runge-Kutta time stepper for the equation eq.

FourierFlows.TimeStepper — Function

TimeStepper(stepper, equation, dt=nothing, dev=CPU(); kw...)

Instantiate the Timestepper for equation with timestep dt and on the device. The kw are passed to the timestepper constructor.

FourierFlows.step_until! — Method

step_until!(prob, stop_time)

Step forward prob until stop_time. Cannot be used with ETDRK4 time steppers.

See also: stepforward!

FourierFlows.stepforward! — Method

stepforward!(prob::Problem, diags, nsteps::Int)

Step forward prob for nsteps, incrementing diags along the way. diags may be a single Diagnostic or a Vector of Diagnostics.

FourierFlows.stepforward! — Method

stepforward!(prob, nsteps::Int)

Step forward prob for nsteps.

FourierFlows.stepforward! — Method

stepforward!(prob)

Step forward prob one time step.

Diffusion Testbed Module

FourierFlows.Diffusion.Problem — Method

Problem(; parameters...)

Construct a constant diffusivity problem.

FourierFlows.Diffusion.set_c! — Method

set_c!(prob, c)

Set the solution as the transform of c and update vars.

FourierFlows.Diffusion.updatevars! — Method

updatevars!(vars, grid, sol)

Update the variables in vars on the grid with the solution in sol.